Drop the smallest zoom level for all view types.
This zoom level is pretty much useless in icon and compact views,
as it shows no filenames.  The icons are also too small to really
be useful either.

In the list view, the size of the icon is smaller than the text
size, which is pointless - the view won't shrink beyond the text
height anyhow, so why would you ever want the icons to be smaller?

This also restores the original icon sizes for the list view (which
were inadvertantly altered when changing the sizes in the icon view
in 82eecad574.  It also updates the zoom percentages to in prefs to
reflect the new actual sizes.